+++ webpack.config.js
@@ -1,0 +1,23 @@
+  function webpackOptions(config){
+    return {
+      webpackOptions: {
+        resolve: {
+          extensions: [".ts", ".js"],
+        },
+        module: {
+          rules: [
+            {
+              test: /\.feature$/,
+              use: [
+                {
+                  loader: "@badeball/cypress-cucumber-preprocessor/webpack",
+                  options: config,
+                },
+              ],
+            },
+          ],
+        },
+      },
+    }
+  }
+module.exports = {webpackOptions}