+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 1.1.0-alpha.1 (2023-11-28)
+* Add new method cucumber-preprocessor use [webpack](https://github.com/cypress-io/cypress/tree/develop/npm/webpack-preprocessor) 
+* Update dependencies 
+* modify structure `module.export` by `export default`
+
 ## 1.0.0 (2023-11-20)